feat(registry): extract content schema helpers into content-schema-lib
Move heading extraction, structured field inference, SEO derivation, and
category validation into a dedicated content-schema-lib module with
comprehensive unit tests so the public content-schema wrapper stays thin.
